Refactored AltQuestionSearch to have search bar and all functionality besides the filtered mapping inside of QandA
